chore: add auxiliary publishing, move metrics into its own package
