+++ packages/metrics/src/index.ts
@@ -1,4 +1,4 @@
 export { PrometheusMetrics } from "./prometheusMetrics.js";
 export { createDefaultMetrics } from "./metricDefinitions.js";
-export type { Metrics, MetricDefinitions } from "./metricsTypes.js";
 export type { DefaultMetrics } from "./metricDefinitions.js";
+export type { Metrics, MetricDefinitions, PrometheusMetricsOptions } from "./types.js";
