+++ src/server.ts
@@ -23,7 +23,7 @@ import { type ConnectionErrorHandler } from "./common/connectionErrorHandler.js"
 import type { Elicitation } from "./elicitation.js";
 import { AllTools } from "./tools/index.js";
 import type { UIRegistry } from "./ui/registry/index.js";
-import type { Metrics, DefaultMetrics } from "./common/metrics/index.js";
+import type { Metrics, DefaultMetrics } from "@mongodb-js/mcp-metrics";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnyToolClass = ToolClass<any, any, any>;
