+++ src/tools/tool.ts
@@ -13,7 +13,7 @@ import type { UIRegistry } from "../ui/registry/index.js";
 import { createUIResource, type UIResource } from "@mcp-ui/server";
 import { TRANSPORT_PAYLOAD_LIMITS, type TransportType } from "../transports/constants.js";
 import { getRandomUUID } from "../helpers/getRandomUUID.js";
-import type { DefaultMetrics, Metrics } from "../common/metrics/index.js";
+import type { Metrics, DefaultMetrics } from "@mongodb-js/mcp-metrics";
 import { redact } from "mongodb-redact";
 
 export type ToolArgs<T extends ZodRawShape> = {
