+++ src/transports/base.ts
@@ -22,8 +22,7 @@ import { applyConfigOverrides } from "../common/config/configOverrides.js";
 import { ApiClient, type ApiClientFactoryFn } from "../common/atlas/apiClient.js";
 import { defaultCreateApiClient } from "../common/atlas/apiClient.js";
 import type { UIRegistry } from "../ui/registry/index.js";
-import { createDefaultMetrics, PrometheusMetrics, type DefaultMetrics } from "../common/metrics/index.js";
-import type { Metrics } from "../common/metrics/metricsTypes.js";
+import { PrometheusMetrics, createDefaultMetrics, type Metrics, type DefaultMetrics } from "@mongodb-js/mcp-metrics";
 
 /**
  * Request context containing HTTP headers and query parameters.
